Move annotations pure functions to annotations module
