+++ assets/js/dashboard/annotations/annotations.ts
@@ -1,3 +1,6 @@
+import { Interval } from "../stats/graph/intervals";
+import { parseUTCDate } from "../util/date";
+
 export enum AnnotationType {
   personal = 'personal',
   site = 'site'
@@ -39,4 +42,59 @@ export type AnnotationPayload = Pick<
 export const ANNOTATION_TYPE_LABELS = {
   [AnnotationType.personal]: 'Personal note',
   [AnnotationType.site]: 'Site-wide note'
+}
+
+export const getAnnotationTimeLabel = (
+  annotation: Pick<Annotation, 'datetime' | 'granularity'>,
+  interval: Interval
+): string => {
+  const dateString = annotation.datetime.substring(0, 'YYYY-MM-DD'.length)
+  switch (annotation.granularity) {
+    case AnnotationGranularity.date: {
+      switch (interval) {
+        case Interval.month:
+          // floors to closest start of month for the date
+          return parseUTCDate(dateString).startOf('month').format('YYYY-MM-DD')
+        case Interval.week:
+          // floors to closest start of week for the date
+          return parseUTCDate(dateString).startOf('week').format('YYYY-MM-DD')
+        case Interval.day:
+        case Interval.hour:
+        case Interval.minute:
+          // floors to date
+          return dateString
+      }
+      break
+    }
+    case AnnotationGranularity.minute: {
+      switch (interval) {
+        case Interval.month:
+          // floors to closest start of month for the date
+          return parseUTCDate(dateString).startOf('month').format('YYYY-MM-DD')
+        case Interval.week:
+          // floors to closest start of week for the date
+          return parseUTCDate(dateString).startOf('week').format('YYYY-MM-DD')
+        case Interval.day:
+          // floors to date
+          return dateString
+        case Interval.hour: {
+          const [dateYYYYMMDD, timeHHMMSS] = annotation.datetime.split('T')
+          // floors time to hour
+          return `${dateYYYYMMDD} ${timeHHMMSS.substring(0, 'HH'.length)}:00:00`
+        }
+        case Interval.minute:
+          return annotation.datetime.split('T').join(' ')
+      }
+    }
+  }
+}
+
+export const groupAnnotationsByTimeLabel = (
+  annotations: Annotation[],
+  interval: Interval
+): Record<string, Annotation[] | undefined> => {
+  return annotations.reduce<Record<string, Annotation[]>>((acc, annotation) => {
+    const timeLabel = getAnnotationTimeLabel(annotation, interval)
+    return { ...acc, [timeLabel]: [...(acc[timeLabel] ?? []), annotation] }
+  }, {})
 }
